Update actuator framework to treats steppers as steppers, not DC motors
Replace LinearActuator with StepperActuator and define (almost
completely) the private methods that StepperActuator needs to define.
Still need to handle initialization and zeroing of initial position, and
the actual mechanics of commanding a single step forward or backward,
but that's all that's left for StepperActuator

Also defines three convenience wrapper classes that are not yet
implemented - each will make use of the StepperActuator API for the
specific kind of control we need for the platform, the nozzle, and the
carriage, and will allow for really simple multithreading.